requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic-settings
reportlab
prometheus-fastapi-instrumentator==7.0.0
weasyprint
jinja2
